[apps/web] channels group as a separate component
